requirements.txt
django
pillow
django-rest-framework
django-crispy-forms